fsp: Move ISaveDataInfoReader to a seperate file
